test(socketio): fix socket.io tests
